requirements.txt
asgiref==3.7.2
Django==4.2.5
django-autoslug==1.9.9
django-bootstrap-v5==1.0.11
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.14
Pillow==10.0.1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
